[MOV] hr_holidays : move code to split file per models
