+++ addons/hr_holidays/__openerp__.py
@@ -33,8 +33,9 @@ A synchronization with an internal agenda (Meetings of the CRM module) is also p
         'security/ir.model.access.csv',
         'security/hr_holidays_security.xml',
 
-        'views/hr_views.xml',
+        'views/ressource_views.xml',
         'views/hr_holidays_views.xml',
+        'views/hr_views.xml',
 
         'wizard/hr_holidays_summary_department_views.xml',
         'wizard/hr_holidays_summary_employees_views.xml',
